# storage/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aphrodite API",
        "description": "API Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Environment"
        },
        {
            "url": "http://test.aphrodite.com",
            "description": "Test Environment"
        }
    ],
    "paths": {
        "/v1/send-code": {
            "post": {
                "tags": [
                    "Auth Moudle"
                ],
                "summary": "Send Verification Code",
                "operationId": "ce471f2d6b74c5b8f92e8f6dd1f2e46c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendVerifyCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "Auth Moudle"
                ],
                "summary": "User Registration/Login",
                "operationId": "f73e99b0f556dcc11b15f2fcac9adf4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Auth Moudle"
                ],
                "summary": "User Logout",
                "operationId": "3b25ddba9ad865a7c2d334aa2e781acb",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/v1/user": {
            "get": {
                "tags": [
                    "User Moudle"
                ],
                "summary": "User Info",
                "operationId": "447f660054fa7c5b26b57e070bbe8f86",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Moudle"
                ],
                "summary": "Update User",
                "operationId": "a60239bfb9bd1ae8a4d9a05b00de7c4d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Moudle"
                ],
                "summary": "Delete User",
                "operationId": "8ffab6d51533fbe237ad9151f90ebe4e",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "phone",
                    "code"
                ],
                "properties": {
                    "phone": {
                        "description": "User Phone",
                        "type": "string",
                        "example": "13800138000"
                    },
                    "code": {
                        "description": "Verification code",
                        "type": "string",
                        "maxLength": 4,
                        "minLength": 4,
                        "example": "1234"
                    }
                },
                "type": "object"
            },
            "SendVerifyCodeRequest": {
                "required": [
                    "phone"
                ],
                "properties": {
                    "phone": {
                        "description": "User Phone",
                        "type": "string",
                        "example": "13800138000"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "nickname": {
                        "description": "User Nickname",
                        "type": "string",
                        "example": "john"
                    },
                    "email": {
                        "description": "User Email",
                        "type": "string",
                        "example": "john@example.com"
                    }
                },
                "type": "object"
            },
            "GetUserResponse": {
                "properties": {
                    "nickname": {
                        "description": "User Nickname",
                        "type": "string",
                        "example": "john"
                    },
                    "userNo": {
                        "description": "User Number",
                        "type": "integer",
                        "example": "A8000"
                    },
                    "userCode": {
                        "description": "User Code",
                        "type": "string",
                        "example": "100000"
                    },
                    "email": {
                        "description": "User Email",
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "description": "User Phone",
                        "type": "string",
                        "example": "13800138000"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "required": [
                    "assessToken"
                ],
                "properties": {
                    "assessToken": {
                        "description": "Access token",
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer Token authentication"
            }
        }
    },
    "tags": [
        {
            "name": "Auth Moudle"
        },
        {
            "name": "User Moudle"
        }
    ]
}